+++ src/components/Hub/AppView.vue
@@ -20,7 +20,7 @@
                 >
                   <template #placeholder>
                     <v-row class="fill-height" align="center" justify="center">
-                      <v-progress-circular indeterminate />
+                      <v-progress-circular indeterminate aria-label="Loading image" />
                     </v-row>
                   </template> </v-img
               ></v-col>