+++ src/components/Hub/HubCard.vue
@@ -53,7 +53,7 @@
             <v-progress-circular
               indeterminate
               color="blue darken-3"
-              aria-label="progress"
+              aria-label="Loading image"
             ></v-progress-circular>
           </v-row>
         </template>
@@ -75,7 +75,7 @@
             <v-progress-circular
               indeterminate
               color="blue darken-3"
-              aria-label="progress"
+              aria-label="Loading image"
             ></v-progress-circular>
           </v-row>
         </template>
@@ -98,7 +98,7 @@
             <v-progress-circular
               indeterminate
               color="blue darken-3"
-              aria-label="progress"
+              aria-label="Loading image"
             ></v-progress-circular>
           </v-row>
         </template>