+++ src/components/____HomeSplash.vue
@@ -91,6 +91,7 @@
                     <v-progress-circular
                       indeterminate
                       color="grey lighten-5"
+                      aria-label="Loading image"
                     ></v-progress-circular>
                   </v-row>
                 </template>